Use pathed archives on Linux
Alpine Linux's `ar` program doesn't support having objects with the
same name in the archive twice, so we have to use the GNU extension
for including paths.

This commit *should* honestly be split into multiple separate commits,
but unfortunately pretty much all of these changes have to be applied
as a unit or a build step fails.